move files into /lib add ./lib to "files" in package.json add command for interactive mode move actions to own file
