+++ lib/loading.js
