+++ package.json
@@ -22,8 +22,7 @@
   },
   "files": [
     "index.js",
-    "az.js",
-    "loading.js"
+    "./lib"
   ],
   "bin": {
     "azx": "./index.js"
